get rid of milestones, refactor for facts
