+++ .env.local.template
@@ -2,7 +2,7 @@ APP_VERSION=$npm_package_version
 APP_NAME=$npm_package_name
 NEXT_PUBLIC_KONTENT_ENVIRONMENT_ID=2cedc519-a547-01ba-dde7-b01e00b909a1
 KONTENT_COLLECTION_CODENAME=ficto_healthtech_surgical
-# Domain for toher collection links in format <COLLECTION_CODENAME>:<DOMAIN_WITOUHOUT_PROTOCOL> i.e. "ficto_healthtech_imaging:ficto-healthtech-imaging.vercel.app,ficto_healthtech_surgical:ficto-healthtech-surgical.vercel.app
+# Domain for other collection links in format <COLLECTION_CODENAME>:<DOMAIN_WITOUHOUT_PROTOCOL> i.e. "ficto_healthtech_imaging:ficto-healthtech-imaging.vercel.app,ficto_healthtech_surgical:ficto-healthtech-surgical.vercel.app
 NEXT_PUBLIC_OTHER_COLLECTIONS_DOMAINS=
 KONTENT_PREVIEW_API_KEY=
 KONTENT_MANAGEMENT_API_KEY=
