+++ models/content-types/Block_VisualContainer.ts
@@ -2,7 +2,6 @@ import { type IContentItem, type Elements } from '@kontent-ai/delivery-sdk';
 import { type Block_EmojiGuideline } from '../content-type-snippets/Block_EmojiGuideline';
 import { type Block_HeroUnit } from './Block_HeroUnit';
 import { type Fact } from './fact';
-import { type Milestone } from './milestone';
 
 /**
  * Generated by '@kontent-ai/model-generator@5.10.0'
@@ -18,7 +17,7 @@ export type Block_VisualContainer = IContentItem<{
    * Id: 081f63c6-9589-5f8d-a9de-bc3b1ceb24e7
    * Codename: items
    */
-  items: Elements.LinkedItemsElement<Fact | Milestone | Block_HeroUnit>;
+  items: Elements.LinkedItemsElement<Fact | Block_HeroUnit>;
 
   /**
    * Subtitle (text)
