+++ models/content-types/WSL_Page.ts
@@ -11,14 +11,14 @@ import { type WSL_WebSpotlightRoot } from './WSL_WebSpotlightRoot';
  * Generated by '@kontent-ai/model-generator@5.10.0'
  *
  * 💡 Page
- * Id: 2d783b04-c9ac-42f5-bd34-959340e8e2ac
+ * Id: 0cb81f69-3f4b-46a1-8ce2-0aada4cdc358
  * Codename: page
  */
 export type WSL_Page = IContentItem<{
   /**
    * Content (modular_content)
    * Required: false
-   * Id: dc3e831e-0157-41c8-b4a4-ff95b576a71e
+   * Id: 27bc37be-796c-4219-a85d-cd61fc2e4f43
    * Codename: content
    */
   content: Elements.LinkedItemsElement<
@@ -28,7 +28,7 @@ export type WSL_Page = IContentItem<{
   /**
    * Subpages (subpages)
    * Required: false
-   * Id: ce96b539-ab2f-4ddf-8e94-e79acf2aec1a
+   * Id: 7e159b22-7b95-4a9b-9e2a-a79d407edd76
    * Codename: subpages
    */
   subpages: Elements.LinkedItemsElement<WSL_Page | WSL_WebSpotlightRoot>;
@@ -36,7 +36,7 @@ export type WSL_Page = IContentItem<{
   /**
    * Title (text)
    * Required: false
-   * Id: aa8bd39c-ad1d-4557-ada1-7422f07a4f47
+   * Id: c2808c7e-9f57-4d0b-815c-f535752ed1cd
    * Codename: title
    */
   title: Elements.TextElement;
@@ -44,7 +44,7 @@ export type WSL_Page = IContentItem<{
   /**
    * URL (url_slug)
    * Required: false
-   * Id: 4e9acc07-f7e9-49ae-a992-88e791b07204
+   * Id: 434f911b-c645-46e1-a633-0c8a3b1e9207
    * Codename: url
    */
   url: Elements.UrlSlugElement;
