+++ models/content-types/WSL_WebSpotlightRoot.ts
@@ -12,14 +12,14 @@ import { type WSL_Page } from './WSL_Page';
  * Generated by '@kontent-ai/model-generator@5.10.0'
  *
  * 💡 Web spotlight root
- * Id: 746e489b-4dbe-403a-a9db-0876fddfb68b
+ * Id: eba96027-914d-48f7-a7f8-38502fa7e869
  * Codename: web_spotlight_root
  */
 export type WSL_WebSpotlightRoot = IContentItem<{
   /**
    * Content (modular_content)
    * Required: false
-   * Id: 376309cf-f65e-43ab-a104-118f5c143413
+   * Id: 0273c87b-86ae-4c9b-8e49-d064f25d1417
    * Codename: content
    */
   content: Elements.LinkedItemsElement<
@@ -37,7 +37,7 @@ export type WSL_WebSpotlightRoot = IContentItem<{
   /**
    * Subpages (subpages)
    * Required: false
-   * Id: a96ae5ad-8971-47ca-9a06-330193ebb7ef
+   * Id: 9e2d4dd9-7d9f-4365-99e0-6b1528f25215
    * Codename: subpages
    */
   subpages: Elements.LinkedItemsElement<WSL_Page | WSL_WebSpotlightRoot>;
@@ -45,7 +45,7 @@ export type WSL_WebSpotlightRoot = IContentItem<{
   /**
    * Title (text)
    * Required: false
-   * Id: 60ff91ab-cd43-4003-a9ba-d50639b57d18
+   * Id: fba9369c-61f3-4a91-a809-70129484c3ca
    * Codename: title
    */
   title: Elements.TextElement;
