+++ models/project/contentTypes.ts
@@ -1,9 +1,9 @@
 /**
  * Generated by '@kontent-ai/model-generator@5.10.0'
  *
- * Project name: New sample migrations (TMP) (Kontent)
- * Environment: App test
- * Project Id: 89962bec-344a-01fd-58aa-40c92b21f579
+ * Project name: Migration test
+ * Environment: Production
+ * Project Id: b5905129-8bdb-0115-aca0-d1700790cd2a
  */
 export const contentTypes = {
   /**
@@ -11,7 +11,7 @@ export const contentTypes = {
    */
   page: {
     codename: 'page',
-    id: '2d783b04-c9ac-42f5-bd34-959340e8e2ac',
+    id: '0cb81f69-3f4b-46a1-8ce2-0aada4cdc358',
     externalId: undefined,
     name: '💡 Page',
     elements: {
@@ -20,7 +20,7 @@ export const contentTypes = {
        */
       content: {
         codename: 'content',
-        id: 'dc3e831e-0157-41c8-b4a4-ff95b576a71e',
+        id: '27bc37be-796c-4219-a85d-cd61fc2e4f43',
         externalId: undefined,
         name: 'Content',
         required: false,
@@ -59,7 +59,7 @@ export const contentTypes = {
        */
       subpages: {
         codename: 'subpages',
-        id: 'ce96b539-ab2f-4ddf-8e94-e79acf2aec1a',
+        id: '7e159b22-7b95-4a9b-9e2a-a79d407edd76',
         externalId: undefined,
         name: 'Subpages',
         required: false,
@@ -72,7 +72,7 @@ export const contentTypes = {
        */
       title: {
         codename: 'title',
-        id: 'aa8bd39c-ad1d-4557-ada1-7422f07a4f47',
+        id: 'c2808c7e-9f57-4d0b-815c-f535752ed1cd',
         externalId: undefined,
         name: 'Title',
         required: false,
@@ -98,7 +98,7 @@ export const contentTypes = {
        */
       url: {
         codename: 'url',
-        id: '4e9acc07-f7e9-49ae-a992-88e791b07204',
+        id: '434f911b-c645-46e1-a633-0c8a3b1e9207',
         externalId: undefined,
         name: 'URL',
         required: false,
@@ -113,7 +113,7 @@ export const contentTypes = {
    */
   web_spotlight_root: {
     codename: 'web_spotlight_root',
-    id: '746e489b-4dbe-403a-a9db-0876fddfb68b',
+    id: 'eba96027-914d-48f7-a7f8-38502fa7e869',
     externalId: undefined,
     name: '💡 Web spotlight root',
     elements: {
@@ -122,7 +122,7 @@ export const contentTypes = {
        */
       content: {
         codename: 'content',
-        id: '376309cf-f65e-43ab-a104-118f5c143413',
+        id: '0273c87b-86ae-4c9b-8e49-d064f25d1417',
         externalId: undefined,
         name: 'Content',
         required: false,
@@ -174,7 +174,7 @@ export const contentTypes = {
        */
       subpages: {
         codename: 'subpages',
-        id: 'a96ae5ad-8971-47ca-9a06-330193ebb7ef',
+        id: '9e2d4dd9-7d9f-4365-99e0-6b1528f25215',
         externalId: undefined,
         name: 'Subpages',
         required: false,
@@ -187,7 +187,7 @@ export const contentTypes = {
        */
       title: {
         codename: 'title',
-        id: '60ff91ab-cd43-4003-a9ba-d50639b57d18',
+        id: 'fba9369c-61f3-4a91-a809-70129484c3ca',
         externalId: undefined,
         name: 'Title',
         required: false,
@@ -680,43 +680,6 @@ export const contentTypes = {
         id: '75c5ed43-3eb7-5902-bf8b-4ddf8f830c8b',
         externalId: 'title-3bd73b39-da80-4ecf-8a53-4972b39df0e4',
         name: 'Title',
-        required: true,
-        type: 'text',
-        snippetCodename: undefined,
-      },
-    },
-  },
-
-  /**
-   * Milestone
-   */
-  milestone: {
-    codename: 'milestone',
-    id: '71d17027-a0e8-5366-8373-cca093fb190d',
-    externalId: 'milestone-7ca0204e-241d-4d77-90a6-e6110232ef14',
-    name: 'Milestone',
-    elements: {
-      /**
-       * Achievement (text)
-       */
-      achievement: {
-        codename: 'achievement',
-        id: '7b5abc87-10de-580b-b350-22e2c1e8c97e',
-        externalId: 'achievement-b5030fb3-1a43-4203-99ab-2368b16f383a',
-        name: 'Achievement',
-        required: true,
-        type: 'text',
-        snippetCodename: undefined,
-      },
-
-      /**
-       * Subject (text)
-       */
-      subject: {
-        codename: 'subject',
-        id: 'd6428ff0-bc13-57df-b7be-70ebe5f79a90',
-        externalId: 'subject-99a49380-1094-4baa-b359-6007624f9869',
-        name: 'Subject',
         required: true,
         type: 'text',
         snippetCodename: undefined,
