+++ models/project/languages.ts
@@ -1,9 +1,9 @@
 /**
  * Generated by '@kontent-ai/model-generator@5.10.0'
  *
- * Project name: New sample migrations (TMP) (Kontent)
- * Environment: App test
- * Project Id: 89962bec-344a-01fd-58aa-40c92b21f579
+ * Project name: Migration test
+ * Environment: Production
+ * Project Id: b5905129-8bdb-0115-aca0-d1700790cd2a
  */
 export const languages = {
   /**
